Update ReadMe Member 1 template
The team name has been created but is subject to change if the collaborator suggests a new one.  The template for member 1 has just been created, also added a comment for instructions on the image size for consistency. an actual code/ syntax line has been typed, the code will work fine only when the path and image source is true.